# project2-branch2/src/openapi.yaml
swagger: "2.0"
info:
  description: ""
  version: "1.0.0"
  title: "Swagger Bookstore"
  # termsOfService: "http://swagger.io/terms/"
  # contact:
  #   email: "apiteam@swagger.io"
  # license:
  #   name: "Apache 2.0"
  #   url: "http://www.apache.org/licenses/LICENSE-2.0.html"
# host: "bookstore.swagger.io"
# basePath: "/v2"
...
schemes:
  - "https"
paths:
...
  /api/v1/publishers/savePublisher:
    post:
      tags:
        - "publisher"
      summary: "Add a new publisher to the store"
      description: ""
      operationId: "savePublisher"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Author object that needs to be added to the store"
          required: true
          schema:
            $ref: "#/definitions/Publisher"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Publisher"
        "405":
          description: "Invalid input"
  /api/v1/publishers/updatePublisher:

    put:
      tags:
        - "publisher"
      summary: "Updates a publisher in the store with form data"
      description: ""
      operationId: "updatePublisher"
      consumes:
        - "application/x-www-form-urlencoded"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Publishers object that needs to be added to the store"
          required: true
          schema:
            $ref: "#/definitions/Publisher"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Publisher"
        "405":
          description: "Invalid input"
...
  /api/v1/publishers/publisher/{publisherID}:
    get:
      tags:
        - "publisher"
      summary: "Find publisher by ID"
      description: "Returns a single publisher"
      operationId: "getPublisherById"
      produces:
        - "application/json"
      parameters:
        - name: "publisherID"
          in: "path"
          description: "ID of publisher to return"
          required: true
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Publisher"
        "404":
          description: "Publisher not found"


    delete:
      tags:
        - "publisher"
      summary: "Deletes a publisher"
      description: ""
      operationId: "deletePublisherById"
      produces:
        - "application/json"
      parameters:
        - name: "api_key"
          in: "header"
          required: false
          type: "string"
        - name: "publisherID"
          in: "path"
          description: "Publisher id to delete"
          required: true
          type: "integer"
          format: "int64"
      responses:
        "404":
          description: "Publisher not found"
  /api/v1/publishers/publisher/publisherName/{publisherName}:
    get:
      tags:
        - "publisher"
      summary: "Get publisher by name"
      description: ""
      operationId: "publisherName"
      produces:
        - "application/xml"
        - "application/json"
      parameters:
        - name: "publisherName"
          in: "path"
          description: "Publisher is returned "
          required: true
          type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Publisher"
        "400":
          description: "Invalid publisherName supplied"
        "404":
          description: "Publisher not found"
...
  /api/v1/users/saveUser:
    post:
      tags:
        - "user"
      summary: "Add a new user to the store"
      description: ""
      operationId: "saveUser"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "User object that needs to be added to the store"
          required: true
          schema:
            $ref: "#/definitions/User"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/User"
        "405":
          description: "Invalid input"
  /api/v1/users/updateUser:

    put:
      tags:
        - "user"
      summary: "Updates a user in the store with form data"
      description: ""
      operationId: "updateUser"
      consumes:
        - "application/x-www-form-urlencoded"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Users object that needs to be added to the store"
          required: true
          schema:
            $ref: "#/definitions/User"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/User"
        "405":
          description: "Invalid input"
...
  /api/v1/users/user/{userID}:
    get:
      tags:
        - "user"
      summary: "Find user by ID"
      description: "Returns a single user"
      operationId: "getUserById"
      produces:
        - "application/json"
      parameters:
        - name: "userID"
          in: "path"
          description: "ID of user to return"
          required: true
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/User"
        "404":
          description: "User not found"


    delete:
      tags:
        - "user"
      summary: "Deletes a user"
      description: ""
      operationId: "deleteUserById"
      produces:
        - "application/json"
      parameters:
        - name: "api_key"
          in: "header"
          required: false
          type: "string"
        - name: "userID"
          in: "path"
          description: "User id to delete"
          required: true
          type: "integer"
          format: "int64"
      responses:
        "404":
          description: "User not found"
  /api/v1/users/user/userName/{userName}:
    get:
      tags:
        - "user"
      summary: "Get user by user name"
      description: ""
      operationId: "userName"
      produces:
        - "application/xml"
        - "application/json"
      parameters:
        - name: "userName"
          in: "path"
          description: "User is returned "
          required: true
          type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/User"
        "400":
          description: "Invalid username supplied"
        "404":
          description: "User not found"
...
definitions:
  Author:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      surname:
        type: "string"
        format: "byte"
      name:
        type: "string"
        format: "byte"
      patronymic:
        type: "string"
        format: "byte"
      dateOfBirth:
        type: "string"
        format: "date"
      bookList:
        type: "array"
        items:
          $ref: "#/definitions/Book"
  Publisher:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
        format: "byte"
      bookList:
        type: "array"
        items:
          $ref: "#/definitions/Book"
  Book:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      price:
        type: "number"
        format: "double"
      authorList:
        type: "array"
        items:
          $ref: "#/definitions/Author"
      genreList:
        type: "array"
        items:
          $ref: "#/definitions/Genre"
      publisher:
        $ref: "#/definitions/Publisher"
      name:
        type: "string"
        format: "byte"
      numOfPages:
        type: "integer"
        format: "int64"
      year:
        type: "string"
        format: "date"
  Genre:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
        format: "byte"
...
  User:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      username:
        type: "string"
        format: "byte"
      password:
        type: "string"
        format: "byte"

      isBlocked:
        type: "boolean"
        format: "byte"

      role:
        type: "string"
        format: "enum"
...